0023303: Expression always true in ShapeFix_Face.cxx
The two corrected if-clause conditions were always true.

In FixSmallAreaWire where the documentation says: "Detects wires with small area (that is less than 100*Precision::PConfusion(). Removes these wires if they are internal." and the bugfix conforms to this.

For 'NeedCheckSplitWire' option in the Perform method the same assumption is made although no documentation on this could be found.