0026657: Data Exchange - IGES/STEP OCAF writers should keep hierarchy and colors when saving non-root elements
Make modification to save in the file part of the XCAF document starting from specified label.
If specified label is component of the high-level assembly then high-level assembly is saved in document with one specified component.
In other case only part of the document starting from the specified label is saved.