0027719: HLRBrep_Algo incorrect output
Intersection of curves and surfaces in HLRBRep is corrected: parametric limits of face are used instead of parametric limits of surface

Building of outlines (Contap_Contour) is corrected: method ComputeCloseLine is now used in all cases.

Correction of HLRBRep_Surface: usage of UVbounds-option

Correction of test cases according to the new behavior.